requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
base64
